requirements.txt
discord.py
aiohttp
Flask
requests
python-dotenv